requirements.txt
coloredlogs==15.0.1
deepmerge==1.1.1
loguru
PyYAML==6.0
pytest
pendulum==3.0.0
apache-airflow==2.2.3